Flask_project/backend-flask/requirements.txt
flask==2.3.2
flask-cors==3.0.10
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0